core: spacedrop: use error enum
