+++ crates/p2p/src/lib.rs
@@ -20,3 +20,18 @@ pub use smart_guards::SmartWriteGuard;
 pub use stream::UnicastStream;
 
 pub use flume;
+
+use thiserror::Error;
+use tokio::sync::{mpsc, oneshot};
+
+#[derive(Debug, Error)]
+pub enum NewStreamError {
+	#[error("No connection methods available for peer")]
+	NoConnectionMethodsAvailable,
+	#[error("The event loop is offline")]
+	EventLoopOffline(mpsc::error::SendError<ConnectionRequest>),
+	#[error("Failed to establish the connection w/ error: {0}")]
+	ConnectionNeverEstablished(oneshot::error::RecvError),
+	#[error("error connecting to peer: {0}")]
+	Connecting(String),
+}
