+++ crates/p2p/src/peer.rs
@@ -4,6 +4,7 @@ use std::{
 	sync::{Arc, PoisonError, RwLock, RwLockReadGuard, RwLockWriteGuard, Weak},
 };
 
+use crate::NewStreamError;
 use thiserror::Error;
 use tokio::sync::{mpsc, oneshot};
 use tracing::warn;
@@ -297,16 +298,4 @@ impl Peer {
 			});
 		}
 	}
-}
-
-#[derive(Debug, Error)]
-pub enum NewStreamError {
-	#[error("No connection methods available for peer")]
-	NoConnectionMethodsAvailable,
-	#[error("The event loop is offline")]
-	EventLoopOffline(mpsc::error::SendError<ConnectionRequest>),
-	#[error("Failed to establish the connection w/ error: {0}")]
-	ConnectionNeverEstablished(oneshot::error::RecvError),
-	#[error("error connecting to peer: {0}")]
-	Connecting(String),
 }
